pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperliquid-ai-bot"
version = "0.1.0"
description = "Hyperliquid-compatible starter bot with safe execution shell, journaling, and web UI"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "OpenAI" }
]
dependencies = [
  "fastapi>=0.111,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "hyperliquid-python-sdk>=0.15.0",
  "eth-account>=0.11,<1.0",
  "python-dotenv>=1.0,<2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
  "ruff>=0.5,<1.0"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 110

[tool.pytest.ini_options]
pythonpath = ["."]
